Prevent incorrect detection of strings
We don't want to detect quote completion after we just inserted the
closing quote. Prevent string closing if the previous character is a
quote.

This will prevent closing an empty string, but that seems like an
uninteresting use case.

Fixes two test cases. One of the fixed cases still fails in vader, but
the usage works in vim.